Melhoria na estrutura de pastas
