+++ index.html
@@ -3,30 +3,31 @@
 <head>
     <meta charset="UTF-8">
     <script type="text/javascript" src="/socket.io/socket.io.js"></script>
-    <script type="text/javascript" src="js/phaser.min.js"></script>
+    <script type="text/javascript" src="js/plugins/phaser.min.js"></script>
     <title>Dominois</title>
 </head>
 <body>
     <div id="game"></div>
-	
-	<script type="text/javascript" src="js/Helpers/StringHelper.js"></script>
-	<script type="text/javascript" src="js/Helpers/AssetsHelper.js"></script>
-	<script type="text/javascript" src="js/Helpers/EventosHelper.js"></script>
-	
-	<script type="text/javascript" src="js/Pedra.js"></script>
-	<script type="text/javascript" src="js/SpritePedra.js"></script>
-	
-	<script type="text/javascript" src="js/Mesa.js"></script>
-	<script type="text/javascript" src="js/PedraFactory.js"></script>
-	
-	<script type="text/javascript" src="js/MesaFactory.js"></script>
-	<script type="text/javascript" src="js/Tela.js"></script>
-	<script type="text/javascript" src="js/SpriteMesa.js"></script>
-	<script type="text/javascript" src="js/MaoPrincipal.js"></script>
-	<script type="text/javascript" src="js/Jogador.js"></script>
-	
-	<script type="text/javascript" src="js/Game.js"></script>
-	<script type="text/javascript" src="js/main.js"></script>
-	<script type="text/javascript" src="js/client.js"></script>
+    
+    <script type="text/javascript" src="js/helpers/StringHelper.js"></script>
+    <script type="text/javascript" src="js/helpers/AssetsHelper.js"></script>
+    <script type="text/javascript" src="js/helpers/EventosHelper.js"></script>
+
+    <script type="text/javascript" src="js/controllers/MaoPrincipal.js"></script>
+    
+    <script type="text/javascript" src="js/models/sprites/SpritePedra.js"></script>
+    <script type="text/javascript" src="js/models/sprites/SpriteMesa.js"></script>
+
+    <script type="text/javascript" src="js/models/Pedra.js"></script>
+    <script type="text/javascript" src="js/models/Jogador.js"></script>
+    <script type="text/javascript" src="js/models/Mesa.js"></script>
+    <script type="text/javascript" src="js/models/Tela.js"></script>
+    <script type="text/javascript" src="js/models/Game.js"></script>
+
+    <script type="text/javascript" src="js/factories/PedraFactory.js"></script>
+    <script type="text/javascript" src="js/factories/MesaFactory.js"></script>
+    
+    <script type="text/javascript" src="js/main.js"></script>
+    <script type="text/javascript" src="js/client.js"></script>
 </body>
 </html>
