--- js/Pedra.js
@@ -1,8 +1,0 @@
-var Pedra = function(nome, valorSuperior, valorInferior, sprite) {
-	var self = this;	
-	this.nome = nome;
-	this.valorSuperior = valorSuperior;
-	this.valorInferior = valorInferior;	
-	
-	this.sprite = sprite;
-}
