--- js/SpriteMesa.js
@@ -1,14 +1,0 @@
-var SpriteMesa = function() {
-	var self = this;	
-	this.nome = "mesa";
-	
-	this.posicao = {
-		x: 50,
-		y: 50
-	}
-	
-	this.tamanho = {
-		altura: 500,
-		largura: 700
-	}
-}
