--- js/SpritePedra.js
@@ -1,7 +1,0 @@
-var SpritePedra = function(nome) {
-	var self = this;
-	
-	this.nome = nome;	
-	this.altura = 130;
-	this.largura = 68;
-}
