+++ js/models/Jogador.js
@@ -1,0 +1,9 @@
+var Jogador = function() {
+    var self = this;
+    this.pedras = [];
+    this.ParaCadaPedra = function(callback) {
+        for (var i = 0; i < self.pedras.length; i++) {
+            callback(self.pedras[i]);
+        }
+    }
+}
