+++ js/models/sprites/SpriteMesa.js
@@ -1,0 +1,14 @@
+var SpriteMesa = function() {
+    var self = this;	
+    this.nome = "mesa";
+    
+    this.posicao = {
+        x: 50,
+        y: 50
+    }
+    
+    this.tamanho = {
+        altura: 500,
+        largura: 700
+    }
+}
