+++ js/models/sprites/SpritePedra.js
@@ -1,0 +1,7 @@
+var SpritePedra = function(nome) {
+    var self = this;
+    
+    this.nome = nome;	
+    this.altura = 130;
+    this.largura = 68;
+}
